fix: simplify the lazy approach
